requirements.txt
groq
edge-tts
sounddevice
soundfile
python-dotenv
gtts
rich
prompt_toolkit
